+++ CHANGELOG.md
@@ -15,6 +15,8 @@
     - `no:adverse-events`
 - **Design Changes**
   - Design link to "View Sheets" now filters sheets using the new search format
+  - Ranges for integers, numerics, and dates are now displayed underneath the
+    variables on the design editor
 - **Site Changes**
   - Improved the display of site members on the site show page on mobile