+++ CHANGELOG.md
@@ -19,6 +19,8 @@
     variables on the design editor
 - **Site Changes**
   - Improved the display of site members on the site show page on mobile
+- **Subject Changes**
+  - Subject file counts are now computed based on cached sheet file counts
 
 ### Bug Fix
 - Variables that exist on multiple designs no longer produce multiple columns in